pyproject.toml
[project]
name = "vllm-huggingface"
version = "0.1.0"
description = "Docker container compatible with HuggingFace Inference Endpoints"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub",
    "python-dotenv",
]
